Types of Hiding Virus
1.	Binder
It is used to bind or collaborate the server and virus that has been created. So if anyone opens the image or songs to get open and start working in background.
2.	Crypter
It basically a tool used for changing binary code of the .exe files and applies many encryptions on it to make it undetectable.
3.	Zipping
In this method you have to put your server and virus and make it in a zip file, it will change extension from .exe to .zip which bypasses the file transfer facility.
{Note: Using WinRAR provides that file with a password and antivirus won’t be able to detect it.}
4.	Extension
It basically changes the extension of the file from .exe or .bat file to .png, .jpg, .jpeg and etc. and then sends it to the user but the user need to rename the extension to .exe file for the virus/server to perform.